Remove app.passages and app.stories
This approach isn't working because a sync() of these collections can have
unexpected results, e.g. if you need to verify a passage's name is unique, then
a sync will clobber any other pending changes. Instead, we add some basic
methods for retrieving models -- these will need to be rewritten for a
server-based backend, but we would have had to do that anyway with the global
collections too.