Fix race condition with drags
What appeared to be happening is that a re-render was triggered on passage item views in the midst of a drag event conclusion.
The previous code consulted the onscreen div's position to decide what to save to the model, so this would cause passages to sometimes
pop back to their original position after a drag. This saves the drag coordinates on the fly to a view property so that a re-render won't
affect the end result.

Possibly resolves #184 but testing is needed.